+++ apps/files/src/components/Collaborators/SharedFilesList.vue
@@ -165,7 +165,7 @@
 import { mapGetters, mapActions } from 'vuex'
 import Mixins from '../../mixins'
 import FileList from '../FileList.vue'
-import NoContentMessage from '../NoContentMessage.vue'
+import NoContentMessage from '../FilesLists/NoContentMessage.vue'
 import SortableColumnHeader from '../FilesLists/SortableColumnHeader.vue'
 import { shareTypes } from '../../helpers/shareTypes'
 import { textUtils } from '../../helpers/textUtils'
