+++ apps/files/src/components/Trashbin.vue
@@ -59,7 +59,7 @@
 import { mapGetters, mapActions } from 'vuex'
 import Mixins from '../mixins'
 import FileList from './FileList.vue'
-import NoContentMessage from './NoContentMessage.vue'
+import NoContentMessage from './FilesLists/NoContentMessage.vue'
 import SortableColumnHeader from './FilesLists/SortableColumnHeader.vue'
 
 export default {
